Versao incio aula

+++ Project1/src/protocolo.h
@@ -27,7 +27,7 @@ typedef struct {
     //Other important variables
     unsigned int bcc2StufSize;
     unsigned int dataStufSize;
-    unsigned int size_of_data_frame;
+    size_t size_of_data_frame;
 } DataStruct;
 
 /**
